WIP : Adicionadas tratativas de erros ao criar campanha, adicionar items, criar sistemas, lojas e adicionar usuarioMestre

+++ Bag-and-Shop-app/Application/DTOs/Store/AddItemOnStoreReqDTO.cs
@@ -5,6 +5,6 @@
         public int StoreId { get; set; }
         public int ItemId { get; set; }
         public int Quantity { get; set; } = 0;
-        public int Price { get; set; } = 0;
+        public double Price { get; set; } = 0;
     }
 }
